requirements.txt
beautifulsoup4==4.8.1
bs4==0.0.1
certifi==2019.9.11
cffi==1.13.1
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.8
Django==2.2.6
django-cors-headers==3.1.1
django-extensions==2.2.5
django-rest-knox==4.1.0
djangorestframework==3.10.3
drf-yasg==1.17.0
fake-headers==1.0.2
html5lib==1.0.1
idna==2.8
inflection==0.3.1
itypes==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
packaging==19.2
Pillow==6.2.0
psycopg2-binary==2.8.3
pycparser==2.19
pyparsing==2.4.2
pytz==2019.3
requests==2.22.0
ruamel.yaml==0.16.5
ruamel.yaml.clib==0.2.0
six==1.12.0
soupsieve==1.9.4
sqlparse==0.3.0
uritemplate==3.0.0
urllib3==1.25.6
webencodings==0.5.1
whitenoise==3.3.1
